show co2 consumption when adding

+++ myclimatepal/myclimatepal/ListItem.swift
@@ -12,10 +12,10 @@ class ListItem: Identifiable {
     let id = UUID()
     let description: String
     let category: String
-    let CO2eqkg: Float
+    let CO2eqkg: Double
     var searchScore: Double = 0
     
-    init(description: String, category: String, CO2eqkg: Float) {
+    init(description: String, category: String, CO2eqkg: Double) {
         self.description = description
         self.category = category
         self.CO2eqkg = CO2eqkg
